BANG-Bang with time variable period 1 or 0

+++ app/other/airbraker/src/n_sensing.hpp
@@ -2,12 +2,12 @@
 namespace NSensing {
 int InitSensors();
 /**
- * @param[out] tempC temperature in celcius from chosen barometer
+ * @param[out] tempC temperature in centi-celcius from chosen barometer
  * @param[out] pressure_Kpa pressure in kilopascals from chosen barometer
  * @param[out] accelMs2 vertical acceleration from chosen IMU (TODO)
  * @param[out] gyroDps 3axis gyroscope reading from chosen IMU
  */
-int MeasureSensors(float &tempC, float &pressureKPa, NTypes::AccelerometerData &accelMs2, NTypes::GyroscopeData &gyroDps);
+int MeasureSensors(int16_t &tempC, float &pressureKPa, NTypes::AccelerometerData &accelMs2, NTypes::GyroscopeData &gyroDps);
 
 
 }; // namespace NSensing
